refactor: rename core-worker
