+++ src/core-worker.ts
